backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.1

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.1

# HTTP client (for health checks)
httpx==0.26.0

# Logging
structlog==24.1.0

# Utilities
croniter==2.0.1
humanize==4.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Security scanning
bandit==1.7.7

# Linting
ruff==0.2.1
